+++ src/click/core.py
@@ -1446,11 +1446,11 @@ class Group(Command):
         ``chain`` is enabled.
     :param kwargs: Other arguments passed to :class:`Command`.
 
-    .. versionchanged:: 8.2
-        Merged with and replaces the ``MultiCommand`` base class.
-
     .. versionchanged:: 8.0
         The ``commands`` argument can be a list of command objects.
+
+    .. versionchanged:: 8.2
+        Merged with and replaces the ``MultiCommand`` base class.
     """
 
     allow_extra_args = True
@@ -2629,7 +2629,9 @@ class Option(Parameter):
         if name is None:
             if not expose_value:
                 return None, opts, secondary_opts
-            raise TypeError("Could not determine name for option")
+            raise TypeError(
+                f"Could not determine name for option with declarations {decls!r}"
+            )
 
         if not opts and not secondary_opts:
             raise TypeError(
@@ -2993,7 +2995,7 @@ class Argument(Parameter):
         if not decls:
             if not expose_value:
                 return None, [], []
-            raise TypeError("Could not determine name for argument")
+            raise TypeError("Argument is marked as exposed, but does not have a name.")
         if len(decls) == 1:
             name = arg = decls[0]
             name = name.replace("-", "_").lower()