+++ src/click/exceptions.py
@@ -6,6 +6,7 @@ from gettext import gettext as _
 from gettext import ngettext
 
 from ._compat import get_text_stderr
+from .globals import resolve_color_default
 from .utils import echo
 from .utils import format_filename
 
@@ -30,6 +31,9 @@ class ClickException(Exception):
 
     def __init__(self, message: str) -> None:
         super().__init__(message)
+        # The context will be removed by the time we print the message, so cache
+        # the color settings here to be used later on (in `show`)
+        self.show_color: bool | None = resolve_color_default()
         self.message = message
 
     def format_message(self) -> str:
@@ -42,7 +46,11 @@ class ClickException(Exception):
         if file is None:
             file = get_text_stderr()
 
-        echo(_("Error: {message}").format(message=self.format_message()), file=file)
+        echo(
+            _("Error: {message}").format(message=self.format_message()),
+            file=file,
+            color=self.show_color,
+        )
 
 
 class UsageError(ClickException):