+++ tests/test_testing.py
@@ -5,6 +5,7 @@ from io import BytesIO
 import pytest
 
 import click
+from click.exceptions import ClickException
 from click.testing import CliRunner
 
 
@@ -197,6 +198,26 @@ def test_with_color():
     result = runner.invoke(cli, color=True)
     assert result.output == f"{click.style('hello world', fg='blue')}\n"
     assert not result.exception
+
+
+def test_with_color_errors():
+    class CLIError(ClickException):
+        def format_message(self) -> str:
+            return click.style(self.message, fg="red")
+
+    @click.command()
+    def cli():
+        raise CLIError("Red error")
+
+    runner = CliRunner()
+
+    result = runner.invoke(cli)
+    assert result.output == "Error: Red error\n"
+    assert result.exception
+
+    result = runner.invoke(cli, color=True)
+    assert result.output == f"Error: {click.style('Red error', fg='red')}\n"
+    assert result.exception
 
 
 def test_with_color_but_pause_not_blocking():