+++ tests/test_utils.py
@@ -36,9 +36,7 @@ def test_echo(runner):
 
 
 def test_echo_custom_file():
-    import io
-
-    f = io.StringIO()
+    f = StringIO()
     click.echo("hello", file=f)
     assert f.getvalue() == "hello\n"
 
@@ -209,7 +207,6 @@ def test_echo_via_pager(monkeypatch, capfd, cat, test):
     assert out == expected_output
 
 
-@pytest.mark.skipif(WIN, reason="Test does not make sense on Windows.")
 def test_echo_color_flag(monkeypatch, capfd):
     isatty = True
     monkeypatch.setattr(click._compat, "isatty", lambda x: isatty)
@@ -232,9 +229,16 @@ def test_echo_color_flag(monkeypatch, capfd):
     assert out == f"{styled_text}\n"
 
     isatty = False
-    click.echo(styled_text)
-    out, err = capfd.readouterr()
-    assert out == f"{text}\n"
+    # Faking isatty() is not enough on Windows;
+    # the implementation caches the colorama wrapped stream
+    # so we have to use a new stream for each test
+    stream = StringIO()
+    click.echo(styled_text, file=stream)
+    assert stream.getvalue() == f"{text}\n"
+
+    stream = StringIO()
+    click.echo(styled_text, file=stream, color=True)
+    assert stream.getvalue() == f"{styled_text}\n"
 
 
 def test_prompt_cast_default(capfd, monkeypatch):