+++ CHANGES.rst
@@ -78,6 +78,8 @@ Unreleased
     :issue:`2705`
 -   Show correct value for flag default when using ``default_map``.
     :issue:`2632`
+-   Fix ``click.echo(color=...)`` passing ``color`` to coloroma so it can be
+    forced on Windows. :issue:`2606`.
 
 
 Version 8.1.7